requirements.txt
chainlit
langchain
langchain-google-genai
google-generativeai
pandas
plotly
sqlalchemy
python-dotenv
tabulate
matplotlib